fix format

+++ services/media_engine/plugins/ffmpeg_adapter/demuxer/ffmpeg_demuxer_plugin.cpp
@@ -2085,7 +2085,8 @@ int Sniff(const std::string& pluginName, std::shared_ptr<DataSource> dataSource)
         MEDIA_LOG_I("effective sniff: dataSize:" PUBLIC_LOG_D32 " " PUBLIC_LOG_S "[" PUBLIC_LOG_D32 "/100]",
             getData, plugin->name, confidence);
     }
-    if ((StartWith(plugin->name, "mp3") && getData < MP3_PROBE_SIZE) || (getData < DEFAULT_READ_SIZE)) { // no data
+    if ((StartWith(plugin->name, "mp3") && static_cast<uint32_t>(getData) < MP3_PROBE_SIZE) ||
+        (static_cast<uint32_t>(getData) < DEFAULT_READ_SIZE)) { // no data
         MEDIA_LOG_I("leak sniff: dataSize:" PUBLIC_LOG_D32 " " PUBLIC_LOG_S "[" PUBLIC_LOG_D32 "/100]",
             getData, plugin->name, confidence);
     }
